R athlete.zones - generalize for all sports and add hr/pace lows
Fixes #4262

+++ src/Python/SIP/Bindings.cpp
@@ -436,7 +436,7 @@ Bindings::athleteZones(PyObject* date, QString sport) const
             indexlow++;
         }
         indexlow=0;
-        foreach(int low, x.pacezoneslow) {
+        foreach(double low, x.pacezoneslow) {
             PyList_SET_ITEM(pacelows, indexlow, PyFloat_FromDouble(low));
             indexlow++;
         }
